Unused variables (#3963)
* Added -Wused-variable flag

* Fixes for clang compiler

* Removed wrong -Wno-error from protobuf compilation

* More fixes

+++ docs/snippets/MULTI3.cpp
@@ -8,7 +8,7 @@ int main() {
     auto cnnNetwork = ie.ReadNetwork("sample.xml");
     std::string allDevices = "MULTI:";
     std::vector<std::string> myriadDevices = ie.GetMetric("MYRIAD", METRIC_KEY(AVAILABLE_DEVICES));
-    for (int i = 0; i < myriadDevices.size(); ++i) {
+    for (size_t i = 0; i < myriadDevices.size(); ++i) {
         allDevices += std::string("MYRIAD.")
                                 + myriadDevices[i]
                                 + std::string(i < (myriadDevices.size() -1) ? "," : "");
